Removed ERROR_COMMITMENT_WITHIN_UNBONDING_PERIOD
Removed the check of whether the buyer was committing to buy within the calculated `UnbondingPeriodLength` (calculated as unbondingRounds * roundLength).

This was excessively constraining, e.g. if a buyer commits 2000 blocks (~7.5 hours) before the end of round N, and the `deliveredByBlock` was 2000 blocks into round N+U-1, then this constraint would have prohibited the order from being committed to.